requirements.txt
streamlit
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
SpeechRecognition
langchain-groq
langchain_core


-e .